% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPartition.R
\name{AllPartition}
\alias{AllPartition}
\title{Create all possible partitions}
\usage{
AllPartition(com_total, nblock)
}
\arguments{
\item{com_total}{The total number of components}

\item{nblock}{The total number of data blocks}
}
\value{
A matrix where each column indicates a specific combination of components
}
\description{
This function creates all possible partitions of component structure
}
