% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSCRsim.R
\name{CSSCRSim}
\alias{CSSCRSim}
\title{The current function simulates data under CSCR model}
\usage{
CSSCRSim(
  ncluster,
  memcluster,
  ncom,
  ndistinct,
  nblock,
  nvar,
  ynvar,
  psparse = 0,
  pnoise_x = 0,
  pnoise_y = 0,
  clust_reg,
  com_rele,
  comp_str = NULL,
  equal_loading = TRUE,
  n_test = NULL,
  test_random = TRUE,
  pmean_x = 0,
  pmean_y = 0,
  p_combase = 0
)
}
\arguments{
\item{ncluster}{a positive integer (>2) that indicates the number of clusters}

\item{memcluster}{a vector of length ncluster where each element indicates the size of each cluster. It can also be a positive integer if all clusters are of equal size.}

\item{ncom}{a positive integer that indicates the number of common components}

\item{ndistinct}{a vector of length nblock where each element indicates the number of distinctive components. It can also be a positive integer if all blocks contain the same amount of distinctive components}

\item{nblock}{a positive integer that indicates the number of predictor blocks}

\item{nvar}{a vector of length nblock where each element indicates the number of variables in each block. It can also be a positive integer if all blocks are of equal size.}

\item{ynvar}{a positive integer that indicates the number of dependent variables}

\item{clust_reg}{a matrix that indicates the cluster-specific regression coefficients associated with each component. Each row indicating the regression coeffiencients of one of the clusters}
}
\description{
The current function simulates data under CSCR model
}
