% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixCenter.R
\name{MatrixCenter}
\alias{MatrixCenter}
\title{Data pre-processing}
\usage{
MatrixCenter(matrix, center, scale)
}
\arguments{
\item{matrix}{the matrix to be processed}

\item{center}{whether column-wise centering is to be carried out}

\item{scale}{whether column-wise standardization is to be carried out}
}
\value{
A processed matrix
}
\description{
Center and standardized each column of a data matrix
}
