% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelect.R
\name{ModelSelect}
\alias{ModelSelect}
\title{Model selection for CSCR}
\usage{
ModelSelect(data_x, data_y, nvar, nblock, r_max, k_max)
}
\arguments{
\item{data_x}{The vector of outcomes}

\item{nvar}{The total number of variables}

\item{nblock}{The total number of data blocks}

\item{r_max}{The maximal number of components}

\item{k_max}{The maximal number of clusters}
}
\value{
A list of the optimal number of clusters, components, the associated loss value, and the index of scree ratio
}
\description{
This function select the optimal number of components and clusters for the existing
Clusterwise Simultaneous Component Regression (CSCR) model.
}
