% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCSCR.R
\name{MultiCSCCR}
\alias{MultiCSCCR}
\title{The current function estimates the CSCR model}
\usage{
MultiCSCCR(
  data_x,
  data_y,
  n_block,
  n_com,
  n_distinct,
  n_var,
  n_cluster,
  alpha = 0.8,
  converge = 1e-08,
  iteration = 1000,
  num_starts = 1,
  type = NULL,
  start_part = NULL
)
}
\description{
The current function estimates the CSCR model
}
