% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairCongruence.R
\name{paircongruence}
\alias{paircongruence}
\title{Calculate the congruence of two list of matrices}
\usage{
paircongruence(lista, listb, cluster_num)
}
\arguments{
\item{lista}{the first list of all matrices}

\item{listb}{the second list of all metrics}

\item{cluster_num}{the number of clusters}
}
\value{
A scale indicating the average congruence level of the two lissts
}
\description{
The function calculates the congruence of two lists as indicated by the average Tucker's congruence
}
