% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreeSelect.R
\name{ScreeSelect}
\alias{ScreeSelect}
\title{Code for model selection with the scree ratio test}
\usage{
ScreeSelect(loss)
}
\arguments{
\item{loss}{A vector that indicates the loss value associated to different levels of model complexity}
}
\value{
The selection of the best point
}
\description{
Select the optimal value of the parameter with a scree ratio test
}
