% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuckerCongruence.R
\name{tuckercongruence}
\alias{tuckercongruence}
\title{matrix similarities}
\usage{
tuckercongruence(matrix1, matrix2)
}
\arguments{
\item{matrix1}{the first matrix}

\item{matrix2}{the second matrix}
}
\value{
A scale indicating the Tucker congruence of the two matrices
}
\description{
calculate matrix similarities as indicated by the Tucker congruence
}
